pyproject.toml
[project]
name = "flatness-guided-hpo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "backpack-for-pytorch>=1.7.1",
    "bohb-hpo>=0.5.2",
    "gpy>=1.13.2",
    "hebo>=0.3.6",
    "keras>=3.0.0",
    "lightning>=2.5.0.post0",
    "numpy>=1.24.4",
    "openml>=0.15.1",
    "opentuner>=0.8.8",
    "ruff>=0.14.0",
    "scikit-learn>=1.6.1",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
]

[tool.ruff]
target-version = "py310"
line-length = 120
exclude = [
    ".git",
    ".ruff_cache",
    ".venv",
    ".vscode",
    "__pycache__"
]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D203",
    "D213",
    "N806",
    "PLR0913",
    "TRY003",
    "COM812",
    "PLE1205"
]
